Only switch to list of user-added data on success.
