+++ lib/Models/addUserCatalogMember.js
@@ -16,7 +16,8 @@ var TerriaError = require('../Core/TerriaError');
  * @param {Boolean} [options.enable=true] True to enable the newly-added member if it is an item; otherwise, false.
  * @param {Boolean} [options.open=true] True to open the newly-added member if it is a group; otherwise, false.
  * @param {Boolean} [options.zoomTo=true] True to zoom and use the clock of the newly-added member if it is an item; otherwise, false.
- * @return {Promise} A promise that resolves when the catalog member has been added successfully.
+ * @return {Promise} A promise that resolves to the catalog item when loaded successfully, or to a {@link TerriaError} if loading fails.  The {@link TerriaError} is reported
+ *                   by raising {@link Terria#error} prior to resolving.  The promise never rejects.
  */
 var addUserCatalogMember = function(terria, newCatalogMemberOrPromise, options) {
     options = defaultValue(options, defaultValue.EMPTY_OBJECT);
@@ -43,6 +44,8 @@ var addUserCatalogMember = function(terria, newCatalogMemberOrPromise, options) 
         }
 
         terria.catalog.userAddedDataGroup.isOpen = true;
+
+        return newCatalogItem;
     }).otherwise(function(e) {
         if (!(e instanceof TerriaError)) {
             e = new TerriaError({
@@ -53,7 +56,7 @@ var addUserCatalogMember = function(terria, newCatalogMemberOrPromise, options) 
 
         terria.error.raiseEvent(e);
 
-        return when.reject(e);
+        return e;
     });
 };
